+++ HexQuickStackStorage/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - ServerSync integration
+- Configuration option to auto stack trophies
 
 ## v1.3.0